small symbol rename and fix some fpc symbols

+++ include/f/f_pc/f_pc_draw_priority.h
@@ -9,7 +9,7 @@ typedef struct draw_priority_class {
 } draw_priority_class;
 
 s32 fpcDwPi_Get(const draw_priority_class* pDwPi);
-void fpcDwPi_Set(draw_priority_class* pDwPi, s32 p);
-void fpcDwPi_Init(draw_priority_class* pDwPi, s32 p);
+void fpcDwPi_Set(draw_priority_class* pDwPi, s16 p);
+void fpcDwPi_Init(draw_priority_class* pDwPi, s16 p);
 
 #endif
